Refactor ctrlPoint function to accept unlimited number of points
Before this commit, the ctrlPoint function only accepted 3 points and didn't
allow this number to change. This is ineffective for finding ctrlPoints for
edge-points: points at the end of the line which do not have points infront.
Now, if we enter two points to the findCtrlPoint function, the control point
will default to being on the original line (which is the desired behavior).